(Oauth) file refactoring
